Disable use of setting groups on IBM i/PASE
On the IBM i platform, setting the process euid always changes the
process effective groups to be the new euid's groups. Also, while
you can call setgid/setgroups/etc the list of groups must be a
subset of the euid's groups; thus you cannot change to the new euid's
groups unless they are already a subset of the current euid's groups,
which is pointless.

With this, key-based auth works, but password-based auth still needs
a custom backend created.